+++ pkg/webui/console/views/device-general-settings/device-general-settings.js
@@ -51,6 +51,7 @@ export default class DeviceGeneralSettings extends React.Component {
     appId: PropTypes.string.isRequired,
     asConfig: PropTypes.stackComponent.isRequired,
     device: PropTypes.device.isRequired,
+    getDefaultMacSettings: PropTypes.func.isRequired,
     isConfig: PropTypes.stackComponent.isRequired,
     jsConfig: PropTypes.stackComponent.isRequired,
     mayEditKeys: PropTypes.bool.isRequired,
@@ -127,7 +128,16 @@ export default class DeviceGeneralSettings extends React.Component {
   }
 
   render() {
-    const { device, isConfig, asConfig, jsConfig, nsConfig, mayEditKeys, mayReadKeys } = this.props
+    const {
+      device,
+      isConfig,
+      asConfig,
+      jsConfig,
+      nsConfig,
+      mayEditKeys,
+      mayReadKeys,
+      getDefaultMacSettings,
+    } = this.props
 
     const isOTAA = isDeviceOTAA(device)
     const { enabled: isEnabled } = isConfig
@@ -216,6 +226,7 @@ export default class DeviceGeneralSettings extends React.Component {
                 onSubmitSuccess={this.handleSubmitSuccess}
                 mayEditKeys={mayEditKeys}
                 mayReadKeys={mayReadKeys}
+                getDefaultMacSettings={getDefaultMacSettings}
               />
             </Collapse>
             <Collapse title={m.asTitle} description={asDescription} disabled={asDisabled}>